requirements.txt
openai-agents
openai-agents[litellm]
python-dotenv
pydantic
pandas
agentops
PyYAML
numpy
matplotlib
seaborn
tqdm
scipy
statsmodels